+++ seedfarmer/__main__.py
@@ -14,16 +14,15 @@
 
 
 import logging
-import os
-from typing import Optional
+from typing import List, Optional
 
 import click
-from dotenv import load_dotenv
 
 import seedfarmer
 from seedfarmer import DEBUG_LOGGING_FORMAT, commands, config, enable_debug
 from seedfarmer.cli_groups import bootstrap, init, list, metadata, projectpolicy, remove, store
 from seedfarmer.output_utils import print_bolded
+from seedfarmer.utils import load_dotenv_files
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -65,8 +64,10 @@ def version() -> None:
 )
 @click.option(
     "--env-file",
-    default=".env",
+    "env_files",
+    default=[".env"],
     help="A relative path to the .env file to load environment variables from",
+    multiple=True,
     required=False,
 )
 @click.option(
@@ -104,24 +105,40 @@ def version() -> None:
     show_default=True,
     type=int,
 )
+@click.option(
+    "--update-seedkit/--no-update-seedkit",
+    default=False,
+    help="Force SeedFarmer to update the SeedKit if found",
+    show_default=True,
+    type=bool,
+)
+@click.option(
+    "--update-project-policy/--no-update-project-policy",
+    default=False,
+    help="Force SeedFarmer to update the deployed Project Policy",
+    show_default=True,
+    type=bool,
+)
 def apply(
     spec: str,
     profile: Optional[str],
     region: Optional[str],
     qualifier: Optional[str],
-    env_file: str,
+    env_files: List[str],
     debug: bool,
     dry_run: bool,
     show_manifest: bool,
     enable_session_timeout: bool,
     session_timeout_interval: int,
+    update_seedkit: bool,
+    update_project_policy: bool,
 ) -> None:
     """Apply manifests to a SeedFarmer managed deployment"""
     if debug:
         enable_debug(format=DEBUG_LOGGING_FORMAT)
 
     # Load environment variables from .env file if it exists
-    load_dotenv(dotenv_path=os.path.join(config.OPS_ROOT, env_file), verbose=True, override=True)
+    load_dotenv_files(config.OPS_ROOT, env_files)
 
     _logger.info("Apply request with manifest %s", spec)
     if dry_run:
@@ -136,6 +153,8 @@ def apply(
         show_manifest=show_manifest,
         enable_session_timeout=enable_session_timeout,
         session_timeout_interval=session_timeout_interval,
+        update_seedkit=update_seedkit,
+        update_project_policy=update_project_policy,
     )
 
 
@@ -177,8 +196,10 @@ def apply(
 )
 @click.option(
     "--env-file",
-    default=".env",
+    "env_files",
+    default=[".env"],
     help="A relative path to the .env file to load environment variables from",
+    multiple=True,
     required=False,
 )
 @click.option(
@@ -208,7 +229,7 @@ def destroy(
     profile: Optional[str],
     region: Optional[str],
     qualifier: Optional[str],
-    env_file: str,
+    env_files: List[str],
     debug: bool,
     enable_session_timeout: bool,
     session_timeout_interval: int,
@@ -218,7 +239,7 @@ def destroy(
         enable_debug(format=DEBUG_LOGGING_FORMAT)
 
     # Load environment variables from .env file if it exists
-    load_dotenv(dotenv_path=os.path.join(config.OPS_ROOT, env_file), verbose=True, override=True)
+    load_dotenv_files(config.OPS_ROOT, env_files)
 
     # MUST use seedfarmer.yaml so we can initialize codeseeder configs
     project = config.PROJECT