+++ seedfarmer/commands/_deployment_commands.py
@@ -145,7 +145,6 @@ def _execute_deploy(
     permissions_boundary_arn: Optional[str] = None,
     codebuild_image: Optional[str] = None,
 ) -> ModuleDeploymentResponse:
-
     parameters = load_parameter_values(
         deployment_name=cast(str, deployment_manifest.name),
         parameters=module_manifest.parameters,
@@ -373,7 +372,9 @@ def _deploy_validated_deployment(
     du.write_deployed_deployment_manifest(deployment_manifest=deployment_manifest)
 
 
-def prime_target_accounts(deployment_manifest: DeploymentManifest) -> None:
+def prime_target_accounts(
+    deployment_manifest: DeploymentManifest, update_seedkit: bool = False, update_project_policy: bool = False
+) -> None:
     _logger.info("Priming Accounts")
     with concurrent.futures.ThreadPoolExecutor(max_workers=len(deployment_manifest.target_accounts_regions)) as workers:
 
@@ -385,8 +386,12 @@ def prime_target_accounts(deployment_manifest: DeploymentManifest) -> None:
 
         params = []
         for target_account_region in deployment_manifest.target_accounts_regions:
-
-            param_d = {"account_id": target_account_region["account_id"], "region": target_account_region["region"]}
+            param_d = {
+                "account_id": target_account_region["account_id"],
+                "region": target_account_region["region"],
+                "update_seedkit": update_seedkit,
+                "update_project_policy": update_project_policy,
+            }
             if target_account_region["network"] is not None:
                 network = commands.load_network_values(
                     cast(NetworkMapping, target_account_region["network"]),
@@ -394,9 +399,9 @@ def prime_target_accounts(deployment_manifest: DeploymentManifest) -> None:
                     target_account_region["account_id"],
                     target_account_region["region"],
                 )
-                param_d["vpc_id"] = network.vpc_id  # type: ignore
-                param_d["private_subnet_ids"] = network.private_subnet_ids  # type: ignore
-                param_d["security_group_ids"] = network.security_group_ids  # type: ignore
+                param_d["vpc_id"] = network.vpc_id
+                param_d["private_subnet_ids"] = network.private_subnet_ids
+                param_d["security_group_ids"] = network.security_group_ids
 
             params.append(param_d)
 
@@ -472,6 +477,7 @@ def destroy_deployment(
                     def _exec_destroy(args: Dict[str, Any]) -> Optional[ModuleDeploymentResponse]:
                         return _execute_destroy(**args)
 
+                    params = []
                     for _module in _group.modules:
                         _process_module_path(module=_module) if _module.path.startswith("git::") else None
 
@@ -662,6 +668,8 @@ def apply(
     show_manifest: bool = False,
     enable_session_timeout: bool = False,
     session_timeout_interval: int = 900,
+    update_seedkit: bool = False,
+    update_project_policy: bool = False,
 ) -> None:
     """
     apply
@@ -695,6 +703,10 @@ def apply(
         If enabled, boto3 Sessions will be reset on the timeout interval
     session_timeout_interval: int
         The interval, in seconds, to reset boto3 Sessions
+    update_seedkit: bool
+        Force update run of seedkit, defaults to False
+    update_project_policy: bool
+        Force update run of managed project policy, defaults to False
 
     Raises
     ------
@@ -752,7 +764,11 @@ def apply(
                 raise seedfarmer.errors.InvalidPathError("Cannot parse manifest file path")
     deployment_manifest.validate_and_set_module_defaults()
 
-    prime_target_accounts(deployment_manifest=deployment_manifest)
+    prime_target_accounts(
+        deployment_manifest=deployment_manifest,
+        update_seedkit=update_seedkit,
+        update_project_policy=update_project_policy,
+    )
 
     module_info_index = du.populate_module_info_index(deployment_manifest=deployment_manifest)
     destroy_manifest = du.filter_deploy_destroy(deployment_manifest, module_info_index)